Revert "Revert "Warnings pt.2 debruijn""
This reverts commit e29485762d13bb0469dc6b9e66be9b642b5dca1e.

+++ assembler/src/debruijn/graph_pack.hpp
@@ -36,7 +36,7 @@ struct graph_pack: private boost::noncopyable {
 
 	size_t k_value;
 
-	graph_t g;
+    graph_t g;
 	index_t index;
 	IdTrackHandler<graph_t> int_ids;
 	EdgesPositionHandler<graph_t> edge_pos;
@@ -45,13 +45,13 @@ struct graph_pack: private boost::noncopyable {
 	Sequence genome;
 	MismatchMasker<graph_t> mismatch_masker;
 
-	explicit graph_pack(size_t k, const std::string &workdir,
-			Sequence const& genome = Sequence(), size_t single_gap = 0,
-			bool careful_labeling = false, bool use_inner_ids = false) :
-			k_value(k), g(k), index(g, k + 1, workdir), int_ids(g,
-					use_inner_ids), edge_pos(g, single_gap, careful_labeling), kmer_mapper(
-					g, k + 1), genome(genome), mismatch_masker(g) {
-	}
+    explicit graph_pack(size_t k, const std::string &workdir,
+            Sequence const& genome = Sequence(), size_t single_gap = 0,
+            bool careful_labeling = false, bool use_inner_ids = false) :
+    k_value(k), g(k), index(g, (unsigned) k + 1, workdir), 
+    int_ids(g, use_inner_ids), edge_pos(g, single_gap, careful_labeling), 
+    kmer_mapper(g, k + 1), genome(genome), mismatch_masker(g) {
+    }
 };
 
 typedef graph_pack<ConjugateDeBruijnGraph, runtime_k::RtSeq,
